Finalized arena mechanism
Currently there is a bug that causes respawned enemies to return tot the (0,0,0) position of the world. It seems that this is caused by the CharacterController probably saving whatever the position of the object was in the previous frame.

I have already tried numerous fixes. But none seem to work, and the bug seems to sometimes occur without ever using the CharacterController as well! I should investigate this more, but should be aware of time constraints as well.